+++ src/main/java/com/tnkj/project/system/line/controller/LineController.java
@@ -1,6 +1,10 @@
 package com.tnkj.project.system.line.controller;
 
 import java.util.List;
+
+import com.tnkj.framework.web.domain.DeptZtree;
+import com.tnkj.project.system.station.domain.Station;
+import com.tnkj.project.system.station.service.IStationService;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -23,7 +27,7 @@ import com.tnkj.framework.web.page.TableDataInfo;
  * 线路Controller
  * 
  * @author tnkj
- * @date 2019-08-19
+ * @date 2019-09-25
  */
 @Controller
 @RequestMapping("/system/line")
@@ -33,6 +37,9 @@ public class LineController extends BaseController
 
     @Autowired
     private ILineService lineService;
+
+    @Autowired
+    private IStationService stationService;
 
     @RequiresPermissions("system:line:view")
     @GetMapping()
@@ -118,8 +125,41 @@ public class LineController extends BaseController
     @Log(title = "线路", businessType = BusinessType.DELETE)
     @PostMapping( "/remove")
     @ResponseBody
-    public AjaxResult remove(String ids)
-    {
+    public AjaxResult remove(String ids){
+        Station station=new Station();
+        station.setLineId(ids);
+        List<Station> stationList=stationService.selectStationList(station);
+        if(stationList!=null && !stationList.isEmpty()){
+            return AjaxResult.warn("线路存在车站,不允许删除");
+        }
         return toAjax(lineService.deleteLineByIds(ids));
+    }
+
+    /**
+     * 加载线路列表树
+     */
+    @GetMapping("/treeData")
+    @ResponseBody
+    public List<DeptZtree> treeData()
+    {
+        List<DeptZtree> ztrees = lineService.selectLineTree(new Line());
+        return ztrees;
+    }
+
+    /**
+     * 选择线路树
+     */
+    @GetMapping("/selectLineTree/{lineId}")
+    public String selectLineTree(@PathVariable("lineId") String lineId, ModelMap mmap)
+    {
+        if("1".equals(lineId)){
+            Line temp=new Line();
+            temp.setId(lineId);
+            temp.setName("郑州电务段");
+            mmap.put("line", temp);
+        }else{
+            mmap.put("line", lineService.selectLineById(lineId));
+        }
+        return prefix + "/tree";
     }
 }
